Fixed Holes

+++ StatsForeGolfers/ProtoWeb/Models/Hole.cs
@@ -9,9 +9,9 @@ namespace ProtoWeb.Models
 {
     public class Hole
     {
-        [Range(typeof(decimal),"1","18")]
+        [Range(typeof(int),"1","18")]
         public int HoleNo { get; set; }
-        [Range(typeof(decimal),"3","5")]
+        [Range(typeof(int),"3","5")]
         public int Par { get; set; }
         public int Length { get; set; }
         public double Handicap { get; set; }
